+++ vulkit/vkit/descriptors/descriptor_pool.cpp
@@ -70,7 +70,7 @@ Result<DescriptorSet> DescriptorPool::Allocate(const VkDescriptorSetLayout p_Lay
 
 void DescriptorPool::Deallocate(const TKit::Span<const VkDescriptorSet> p_Sets) const noexcept
 {
-    vkFreeDescriptorSets(m_Device, m_Pool, static_cast<u32>(p_Sets.size()), p_Sets.data());
+    vkFreeDescriptorSets(m_Device, m_Pool, p_Sets.size(), p_Sets.data());
 }
 
 void DescriptorPool::Deallocate(const VkDescriptorSet p_Set) const noexcept
